+++ libflv/source/mpeg4-mp4toannexb.c
@@ -22,7 +22,7 @@ static int h264_sps_pps_size(const struct mpeg4_avc_t* avc)
 	return n;
 }
 
-int h264_mp4toannexb(const struct mpeg4_avc_t* avc, const void* data, int bytes, void* out, int size)
+int h264_mp4toannexb(const struct mpeg4_avc_t* avc, const void* data, size_t bytes, void* out, size_t size)
 {
 	int i, n;
 	uint8_t sps_pps_flag;
